+++ BCHWS-v1.0/src/bch/ws/interfaces/CRUD.java
@@ -2,9 +2,12 @@ package bch.ws.interfaces;
 
 import java.util.List;
 
+import bch.hb.mappings.User;
+
 public interface CRUD {
 	public List<?> retrieveAll();
-	public Object retrieveById(int id);
+	public Object retrieveById(long id);
 	public boolean updateRecord(Object obj);
 	public boolean createRecord(Object obj);
+	
 }
